Use 16-bit ints instead of 32-bit ints
The data is 12-bits anyways so why store it as a bigger int?

Note that this didn't work well until the DAC MemDataAlignment was set to DMA_MDATAALIGN_HALFWORD.  It would play the whole clip twice as fast and then continue playing at normal speed.

I suspect it was reading 2 16-bit ints every time instead of 1 32-bit int, and because in a 32-bit register only right 12-bits are read, 1 of the 16-bit ints was unused.

+++ audio_test_2/Core/Src/main.c
@@ -52,7 +52,7 @@ TIM_HandleTypeDef htim2;
 /* USER CODE BEGIN PV */
 // Sample rate: 16000 Hz
 // Number of samples: 40124
-const uint32_t audio_clip_1[] = {
+const uint16_t audio_clip_1[] = {
   2055,   2055,   2055,   2055,   2055,   2039,   2055,   2039,
   2055,   2055,   2055,   2039,   2055,   2039,   2055,   2055,
   2039,   2055,   2055,   2055,   2055,   2055,   2039,   2055,
@@ -5071,7 +5071,7 @@ const uint32_t audio_clip_1[] = {
   2039,   2055,   2039,   2055,
 };
 
-const uint32_t audio_clip_1_length = sizeof(audio_clip_1) / sizeof(audio_clip_1[0]);
+const uint16_t audio_clip_1_length = sizeof(audio_clip_1) / sizeof(audio_clip_1[0]);
 
 
 /* USER CODE END PV */
